+++ apps/portfolio/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @alvarosabu/portfolio
 
+## 1.3.0
+
+### Minor Changes
+
+- df83d289: Home Hero with Pancake Planet
+
+## 1.2.0
+
+### Minor Changes
+
+- a3816362: Youube portfolio section
+
 ## 1.1.1
 
 ### Patch Changes